autodoc/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
docker==7.1.0
python-nmap==0.7.1
paramiko==3.5.0
apscheduler==3.10.4
